// src/app/pages/principal/interfaces/starship.interface.ts
export interface StarShip {
  id?: number;
  name?: string;
  model?: string;
  manufacturer?: string;
  passengers?: string;
  length?: string;
  created?: string;
  imageUrl?: string;
  imageSrcBlob?: string;
  url?: string;
  MGLT?: string;
  cargo_capacity?: string;
  consumables?: string;
  crew?: string;
  edited?: string;
  films?: string[];
  cost_in_credits?: string;
  hyperdrive_rating?: string;
  max_atmosphering_speed?: string;
  pilots?: string[];
  starship_class?: string;
}